Refactor core/bench modules and remove duplicate internals
