+++ src/cauchy_generator/bench/micro.py
@@ -7,6 +7,21 @@ from typing import Callable
 
 import torch
 
+from cauchy_generator.bench.constants import (
+    MICROBENCH_BASE_SEED_OFFSET,
+    MICROBENCH_DEFAULT_REPEATS,
+    MICROBENCH_GENERATE_ONE_SEED_OFFSET,
+    MICROBENCH_LINEAR_SEED,
+    MICROBENCH_NODE_PIPELINE_SEED,
+    MICROBENCH_PARENT_DIM,
+    MICROBENCH_SYNTH_FEATURES,
+    MICROBENCH_SYNTH_ROWS,
+    MICRO_CONFIG_N_FEATURES_CAP,
+    MICRO_CONFIG_N_NODES_CAP,
+    MICRO_CONFIG_N_TEST_CAP,
+    MICRO_CONFIG_N_TRAIN_CAP,
+    MILLISECONDS_PER_SECOND,
+)
 from cauchy_generator.config import GeneratorConfig
 from cauchy_generator.core.dataset import generate_one
 from cauchy_generator.core.node_pipeline import ConverterSpec, apply_node_pipeline
@@ -20,19 +35,19 @@ def _time_ms(func: Callable[[], None], repeats: int) -> float:
     for _ in range(max(1, repeats)):
         func()
     elapsed = time.perf_counter() - start
-    return (elapsed / max(1, repeats)) * 1000.0
+    return (elapsed / max(1, repeats)) * MILLISECONDS_PER_SECOND
 
 
 def _micro_config(config: GeneratorConfig) -> GeneratorConfig:
     """Create a lightweight config variant for inexpensive microbenchmarks."""
 
     c = GeneratorConfig.from_dict(config.to_dict())
-    c.dataset.n_train = min(128, c.dataset.n_train)
-    c.dataset.n_test = min(64, c.dataset.n_test)
-    c.dataset.n_features_min = min(16, c.dataset.n_features_min)
-    c.dataset.n_features_max = min(16, c.dataset.n_features_max)
-    c.graph.n_nodes_min = min(8, c.graph.n_nodes_min)
-    c.graph.n_nodes_max = min(8, c.graph.n_nodes_max)
+    c.dataset.n_train = min(MICRO_CONFIG_N_TRAIN_CAP, c.dataset.n_train)
+    c.dataset.n_test = min(MICRO_CONFIG_N_TEST_CAP, c.dataset.n_test)
+    c.dataset.n_features_min = min(MICRO_CONFIG_N_FEATURES_CAP, c.dataset.n_features_min)
+    c.dataset.n_features_max = min(MICRO_CONFIG_N_FEATURES_CAP, c.dataset.n_features_max)
+    c.graph.n_nodes_min = min(MICRO_CONFIG_N_NODES_CAP, c.graph.n_nodes_min)
+    c.graph.n_nodes_max = min(MICRO_CONFIG_N_NODES_CAP, c.graph.n_nodes_max)
     return c
 
 
@@ -40,22 +55,27 @@ def run_microbenchmarks(
     config: GeneratorConfig,
     *,
     device: str | None = None,
-    repeats: int = 5,
+    repeats: int = MICROBENCH_DEFAULT_REPEATS,
 ) -> dict[str, float | int]:
     """Run targeted microbenchmarks for core generation components."""
 
     g_rng = torch.Generator(device="cpu")
-    g_rng.manual_seed(config.seed + 41)
-    x = torch.randn(512, 64, generator=g_rng)
+    g_rng.manual_seed(config.seed + MICROBENCH_BASE_SEED_OFFSET)
+    x = torch.randn(MICROBENCH_SYNTH_ROWS, MICROBENCH_SYNTH_FEATURES, generator=g_rng)
 
     def run_linear() -> None:
         local_g = torch.Generator(device="cpu")
-        local_g.manual_seed(1)
-        _ = apply_random_function(x, local_g, out_dim=64, function_type="linear")
+        local_g.manual_seed(MICROBENCH_LINEAR_SEED)
+        _ = apply_random_function(
+            x,
+            local_g,
+            out_dim=MICROBENCH_SYNTH_FEATURES,
+            function_type="linear",
+        )
 
     parent_data = [
-        torch.randn(512, 24),
-        torch.randn(512, 24),
+        torch.randn(MICROBENCH_SYNTH_ROWS, MICROBENCH_PARENT_DIM),
+        torch.randn(MICROBENCH_SYNTH_ROWS, MICROBENCH_PARENT_DIM),
     ]
     specs = [
         ConverterSpec(key="feature_0", kind="num", dim=1),
@@ -65,13 +85,17 @@ def run_microbenchmarks(
 
     def run_node_pipeline() -> None:
         g = torch.Generator(device="cpu")
-        g.manual_seed(2)
-        _ = apply_node_pipeline(parent_data, 512, specs, g, "cpu")
+        g.manual_seed(MICROBENCH_NODE_PIPELINE_SEED)
+        _ = apply_node_pipeline(parent_data, MICROBENCH_SYNTH_ROWS, specs, g, "cpu")
 
     micro_cfg = _micro_config(config)
 
     def run_generate_one() -> None:
-        _ = generate_one(micro_cfg, seed=micro_cfg.seed + 99, device=device)
+        _ = generate_one(
+            micro_cfg,
+            seed=micro_cfg.seed + MICROBENCH_GENERATE_ONE_SEED_OFFSET,
+            device=device,
+        )
 
     return {
         "micro_repeats": int(max(1, repeats)),
