+++ src/cauchy_generator/bench/suite.py
@@ -4,36 +4,63 @@ from __future__ import annotations
 
 import datetime as dt
 import hashlib
-import math
-import pickle
 import copy
 import re
 import resource
-import statistics
 import sys
-import tempfile
 import time
 from collections import Counter
-from collections.abc import Callable, Iterable
+from collections.abc import Iterable
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
 import torch
+import cauchy_generator.bench.guardrails as _guardrails
 from cauchy_generator.bench.baseline import compare_summary_to_baseline
+from cauchy_generator.bench.constants import (
+    CURRICULUM_GUARDRAIL_RUNTIME_GATING_MIN_SAMPLE,
+    DIAGNOSTICS_DUPLICATE_PROFILE_SUFFIX_BASE,
+    KIB,
+    LATENCY_SEED_OFFSET,
+    MIB,
+    MICROBENCH_REPEATS,
+    MISSINGNESS_RATE_FAIL_ABS_ERROR,
+    MISSINGNESS_RATE_WARN_ABS_ERROR,
+    PROFILE_KEY_HASH_SUFFIX_LEN,
+    REPRODUCIBILITY_SEED_OFFSET,
+    SMOKE_LATENCY_SAMPLES_CAP,
+    SMOKE_N_FEATURES_CAP,
+    SMOKE_N_NODES_CAP,
+    SMOKE_N_TEST_CAP,
+    SMOKE_N_TRAIN_CAP,
+    SMOKE_NUM_DATASETS_CAP,
+    SMOKE_WARMUP_DATASETS_CAP,
+)
 from cauchy_generator.bench.micro import run_microbenchmarks
 from cauchy_generator.bench.metrics import (
     degradation_percent,
     reproducibility_signature,
     summarize_latencies,
 )
+from cauchy_generator.bench.collectors import (
+    _CurriculumMetadataCollector,
+    _MissingnessAcceptanceCollector,
+    _compose_bundle_callback,
+)
+from cauchy_generator.bench.guardrails import (
+    _build_guardrail_issue,
+    _collect_guardrail_regression_issues,
+    _issue_sort_key,
+    _resolve_curriculum_guardrail_context,
+    _severity_from_thresholds,
+    _status_from_issues,
+)
 from cauchy_generator.bench.throughput import run_throughput_benchmark
 from cauchy_generator.config import (
-    CURRICULUM_STAGE_AUTO,
     CURRICULUM_STAGE_OFF,
     GeneratorConfig,
     MISSINGNESS_MECHANISM_NONE,
-    normalize_curriculum_stage,
 )
 from cauchy_generator.core.dataset import generate_batch_iter, generate_one
 from cauchy_generator.diagnostics import (
@@ -42,13 +69,15 @@ from cauchy_generator.diagnostics import (
     write_coverage_summary_json,
     write_coverage_summary_markdown,
 )
-from cauchy_generator.io.parquet_writer import write_parquet_shards_stream
 from cauchy_generator.hardware import (
     HardwareInfo,
     apply_hardware_profile,
     detect_hardware,
 )
-from cauchy_generator.math_utils import to_numpy as _to_numpy
+from cauchy_generator.meta_targets import (
+    coerce_quantiles as _coerce_quantiles_shared,
+    merge_target_bands,
+)
 from cauchy_generator.rng import SeedManager
 from cauchy_generator.types import DatasetBundle
 
@@ -58,13 +87,6 @@ DEFAULT_PROFILE_CONFIGS: dict[str, str] = {
     "cuda_desktop": "configs/benchmark_cuda_desktop.yaml",
     "cuda_h100": "configs/benchmark_cuda_h100.yaml",
 }
-MISSINGNESS_RATE_WARN_ABS_ERROR = 0.03
-MISSINGNESS_RATE_FAIL_ABS_ERROR = 0.05
-LINEAGE_GUARDRAIL_SMOKE_SAMPLE_CAP = 5
-LINEAGE_GUARDRAIL_SEED_OFFSET = 123_000
-LINEAGE_GUARDRAIL_RUNTIME_TRIALS = 3
-LINEAGE_GUARDRAIL_RUNTIME_GATING_MIN_SAMPLE = 5
-CURRICULUM_GUARDRAIL_RUNTIME_GATING_MIN_SAMPLE = 5
 
 
 @dataclass(slots=True)
@@ -93,8 +115,8 @@ def _peak_rss_mb() -> float:
 
     rss = float(resource.getrusage(resource.RUSAGE_SELF).ru_maxrss)
     if sys.platform == "darwin":
-        return rss / (1024.0 * 1024.0)
-    return rss / 1024.0
+        return rss / MIB
+    return rss / KIB
 
 
 def _profile_counts(
@@ -117,8 +139,8 @@ def _profile_counts(
         warmup = int(warmup_override)
 
     if suite == "smoke":
-        num = min(num, 25)
-        warmup = min(warmup, 5)
+        num = min(num, SMOKE_NUM_DATASETS_CAP)
+        warmup = min(warmup, SMOKE_WARMUP_DATASETS_CAP)
 
     return max(1, num), max(0, warmup)
 
@@ -128,7 +150,7 @@ def _latency_sample_count(config: GeneratorConfig, suite: str, num_datasets: int
 
     n = max(1, min(int(config.benchmark.latency_num_samples), num_datasets))
     if suite == "smoke":
-        return min(n, 5)
+        return min(n, SMOKE_LATENCY_SAMPLES_CAP)
     return n
 
 
@@ -140,7 +162,7 @@ def _collect_latency(
 ) -> dict[str, float]:
     """Collect per-dataset latency samples by repeatedly calling ``generate_one``."""
 
-    manager = SeedManager(config.seed + 77_000)
+    manager = SeedManager(config.seed + LATENCY_SEED_OFFSET)
     samples: list[float] = []
     for i in range(max(1, num_samples)):
         seed = manager.child("latency", i)
@@ -159,7 +181,7 @@ def _collect_reproducibility(
     """Generate two deterministic runs and compare content digests."""
 
     n = max(1, num_datasets)
-    run_seed = config.seed + 91_000
+    run_seed = config.seed + REPRODUCIBILITY_SEED_OFFSET
     sig_a = reproducibility_signature(
         generate_batch_iter(config, num_datasets=n, seed=run_seed, device=device)
     )
@@ -190,60 +212,26 @@ def _prepare_config_for_profile(
     cfg = apply_hardware_profile(cfg, hw)
 
     if suite == "smoke":
-        cfg.dataset.n_train = min(cfg.dataset.n_train, 256)
-        cfg.dataset.n_test = min(cfg.dataset.n_test, 128)
-        cfg.dataset.n_features_min = min(cfg.dataset.n_features_min, 24)
-        cfg.dataset.n_features_max = min(cfg.dataset.n_features_max, 24)
-        cfg.graph.n_nodes_min = min(cfg.graph.n_nodes_min, 16)
-        cfg.graph.n_nodes_max = min(cfg.graph.n_nodes_max, 16)
+        cfg.dataset.n_train = min(cfg.dataset.n_train, SMOKE_N_TRAIN_CAP)
+        cfg.dataset.n_test = min(cfg.dataset.n_test, SMOKE_N_TEST_CAP)
+        cfg.dataset.n_features_min = min(cfg.dataset.n_features_min, SMOKE_N_FEATURES_CAP)
+        cfg.dataset.n_features_max = min(cfg.dataset.n_features_max, SMOKE_N_FEATURES_CAP)
+        cfg.graph.n_nodes_min = min(cfg.graph.n_nodes_min, SMOKE_N_NODES_CAP)
+        cfg.graph.n_nodes_max = min(cfg.graph.n_nodes_max, SMOKE_N_NODES_CAP)
 
     return cfg, requested_device, hw
-
-
-def _coerce_target_bands(raw: object) -> dict[str, tuple[float, float]]:
-    """Normalize target band mappings into finite `(lo, hi)` tuples."""
-
-    normalized: dict[str, tuple[float, float]] = {}
-    if not isinstance(raw, dict):
-        return normalized
-    for key, value in raw.items():
-        if not isinstance(key, str):
-            continue
-        if not isinstance(value, (list, tuple)) or len(value) < 2:
-            continue
-        try:
-            lo = float(value[0])
-            hi = float(value[1])
-        except (TypeError, ValueError):
-            continue
-        if not (math.isfinite(lo) and math.isfinite(hi)):
-            continue
-        normalized[key] = (lo, hi) if lo <= hi else (hi, lo)
-    return normalized
 
 
 def _resolve_target_bands(config: GeneratorConfig) -> dict[str, tuple[float, float]]:
     """Merge diagnostics and top-level target mappings for coverage aggregation."""
 
-    merged = _coerce_target_bands(config.diagnostics.meta_feature_targets)
-    merged.update(_coerce_target_bands(config.meta_feature_targets))
-    return merged
+    return merge_target_bands(config.diagnostics.meta_feature_targets, config.meta_feature_targets)
 
 
 def _coerce_quantiles(raw: object) -> tuple[float, ...]:
     """Normalize diagnostics quantiles into finite values."""
 
-    if not isinstance(raw, (list, tuple)):
-        return ()
-    normalized: list[float] = []
-    for item in raw:
-        try:
-            value = float(item)
-        except (TypeError, ValueError):
-            continue
-        if math.isfinite(value):
-            normalized.append(value)
-    return tuple(normalized)
+    return _coerce_quantiles_shared(raw)
 
 
 def _sanitize_profile_key(profile_key: str) -> str:
@@ -252,7 +240,7 @@ def _sanitize_profile_key(profile_key: str) -> str:
     normalized = re.sub(r"[^A-Za-z0-9._-]+", "_", profile_key).strip("._-")
     if not normalized:
         normalized = "profile"
-    suffix = hashlib.sha1(profile_key.encode("utf-8")).hexdigest()[:8]
+    suffix = hashlib.sha1(profile_key.encode("utf-8")).hexdigest()[:PROFILE_KEY_HASH_SUFFIX_LEN]
     return f"{normalized}_{suffix}"
 
 
@@ -287,27 +275,17 @@ def _is_missingness_enabled(config: GeneratorConfig) -> bool:
 
 
 def _lineage_guardrail_sample_count(config: GeneratorConfig, suite: str, num_datasets: int) -> int:
-    """Choose dataset count used for lineage overhead guardrail checks."""
+    """Compatibility wrapper for suite-level monkeypatching in tests."""
 
-    n = max(1, min(int(config.benchmark.latency_num_samples), num_datasets))
-    if suite == "smoke":
-        return min(n, LINEAGE_GUARDRAIL_SMOKE_SAMPLE_CAP)
-    return n
+    return _guardrails._lineage_guardrail_sample_count(
+        config, suite=suite, num_datasets=num_datasets
+    )
 
 
 def _bundle_without_lineage(bundle: DatasetBundle) -> DatasetBundle:
-    """Clone one bundle while removing lineage metadata for control persistence runs."""
+    """Compatibility wrapper for suite-level monkeypatching in tests."""
 
-    metadata = dict(bundle.metadata)
-    metadata.pop("lineage", None)
-    return DatasetBundle(
-        X_train=bundle.X_train,
-        y_train=bundle.y_train,
-        X_test=bundle.X_test,
-        y_test=bundle.y_test,
-        feature_types=list(bundle.feature_types),
-        metadata=metadata,
-    )
+    return _guardrails._bundle_without_lineage(bundle)
 
 
 def _measure_persistence_datasets_per_minute(
@@ -316,69 +294,25 @@ def _measure_persistence_datasets_per_minute(
     config: GeneratorConfig,
     num_bundles: int,
 ) -> float:
-    """Measure write-path throughput for a fixed number of streamed bundles."""
+    """Compatibility wrapper for suite-level monkeypatching in tests."""
 
-    if num_bundles <= 0:
-        return 0.0
-
-    start = time.perf_counter()
-    with tempfile.TemporaryDirectory(prefix="cauchy_lineage_guardrail_") as tmp_dir:
-        out_dir = Path(tmp_dir) / "shards"
-        _ = write_parquet_shards_stream(
-            bundles,
-            out_dir=out_dir,
-            shard_size=max(1, int(config.output.shard_size)),
-            compression=str(config.output.compression),
-        )
-    elapsed = time.perf_counter() - start
-    if elapsed <= 0.0:
-        return 0.0
-    return (float(num_bundles) / elapsed) * 60.0
+    return _guardrails._measure_persistence_datasets_per_minute(
+        bundles,
+        config=config,
+        num_bundles=num_bundles,
+    )
 
 
 def _stage_bundle(path: Path, bundle: DatasetBundle, *, strip_lineage: bool) -> None:
-    """Serialize one bundle to disk for later replay in persistence timing."""
+    """Compatibility wrapper for suite-level monkeypatching in tests."""
 
-    staged_bundle = _bundle_without_lineage(bundle) if strip_lineage else bundle
-    payload: dict[str, Any] = {
-        "X_train": _to_numpy(staged_bundle.X_train),
-        "y_train": _to_numpy(staged_bundle.y_train),
-        "X_test": _to_numpy(staged_bundle.X_test),
-        "y_test": _to_numpy(staged_bundle.y_test),
-        "feature_types": list(staged_bundle.feature_types),
-        "metadata": dict(staged_bundle.metadata),
-    }
-    with path.open("wb") as f:
-        pickle.dump(payload, f, protocol=pickle.HIGHEST_PROTOCOL)
+    _guardrails._stage_bundle(path, bundle, strip_lineage=strip_lineage)
 
 
 def _iter_staged_bundles(stage_dir: Path, *, num_bundles: int) -> Iterable[DatasetBundle]:
-    """Replay staged bundles from disk as a streaming iterator."""
+    """Compatibility wrapper for suite-level monkeypatching in tests."""
 
-    for idx in range(max(0, num_bundles)):
-        bundle_path = stage_dir / f"bundle_{idx:06d}.pkl"
-        with bundle_path.open("rb") as f:
-            payload = pickle.load(f)
-        if not isinstance(payload, dict):
-            raise ValueError(f"Invalid staged bundle payload at {bundle_path}.")
-
-        feature_types_raw = payload.get("feature_types")
-        metadata_raw = payload.get("metadata")
-        if not isinstance(feature_types_raw, list):
-            raise ValueError(f"Staged feature_types must be a list at {bundle_path}.")
-        if not isinstance(metadata_raw, dict):
-            raise ValueError(f"Staged metadata must be an object at {bundle_path}.")
-        feature_types = [str(value) for value in feature_types_raw]
-        metadata = dict(metadata_raw)
-
-        yield DatasetBundle(
-            X_train=payload["X_train"],
-            y_train=payload["y_train"],
-            X_test=payload["X_test"],
-            y_test=payload["y_test"],
-            feature_types=feature_types,
-            metadata=metadata,
-        )
+    return _guardrails._iter_staged_bundles(stage_dir, num_bundles=num_bundles)
 
 
 def _stage_lineage_trial_bundles(
@@ -390,29 +324,18 @@ def _stage_lineage_trial_bundles(
     baseline_stage_dir: Path,
     current_stage_dir: Path,
 ) -> tuple[int, int]:
-    """Generate and stage baseline/current trial bundles once outside timed sections."""
+    """Compatibility wrapper preserving suite-level generate/serialize patch points."""
 
-    baseline_stage_dir.mkdir(parents=True, exist_ok=True)
-    current_stage_dir.mkdir(parents=True, exist_ok=True)
-
-    seen = 0
-    with_lineage = 0
-    for idx, bundle in enumerate(
-        generate_batch_iter(
-            config,
-            num_datasets=sample_n,
-            seed=seed,
-            device=device,
-        )
-    ):
-        has_lineage = isinstance(bundle.metadata.get("lineage"), dict)
-        if has_lineage:
-            with_lineage += 1
-        file_name = f"bundle_{idx:06d}.pkl"
-        _stage_bundle(current_stage_dir / file_name, bundle, strip_lineage=False)
-        _stage_bundle(baseline_stage_dir / file_name, bundle, strip_lineage=True)
-        seen = idx + 1
-    return with_lineage, seen
+    return _guardrails._stage_lineage_trial_bundles_impl(
+        config,
+        sample_n=sample_n,
+        seed=seed,
+        device=device,
+        baseline_stage_dir=baseline_stage_dir,
+        current_stage_dir=current_stage_dir,
+        generate_batch_iter_fn=generate_batch_iter,
+        stage_bundle_fn=_stage_bundle,
+    )
 
 
 def _measure_lineage_persistence_trials(
@@ -423,59 +346,23 @@ def _measure_lineage_persistence_trials(
     config: GeneratorConfig,
     trials: int,
 ) -> tuple[list[float], list[float]]:
-    """Collect paired baseline/lineage persistence throughput trials."""
+    """Compatibility wrapper preserving suite-level replay/measure patch points."""
 
-    n_trials = max(1, int(trials))
-    baseline_trials: list[float] = []
-    current_trials: list[float] = []
-    for _ in range(n_trials):
-        baseline_bundles = _iter_staged_bundles(baseline_stage_dir, num_bundles=num_bundles)
-        baseline_trials.append(
-            _measure_persistence_datasets_per_minute(
-                baseline_bundles,
-                config=config,
-                num_bundles=num_bundles,
-            )
-        )
-        current_bundles = _iter_staged_bundles(current_stage_dir, num_bundles=num_bundles)
-        current_trials.append(
-            _measure_persistence_datasets_per_minute(
-                current_bundles,
-                config=config,
-                num_bundles=num_bundles,
-            )
-        )
-    return baseline_trials, current_trials
+    return _guardrails._measure_lineage_persistence_trials_impl(
+        baseline_stage_dir=baseline_stage_dir,
+        current_stage_dir=current_stage_dir,
+        num_bundles=num_bundles,
+        config=config,
+        trials=trials,
+        iter_staged_bundles_fn=_iter_staged_bundles,
+        measure_persistence_fn=_measure_persistence_datasets_per_minute,
+    )
 
 
 def _median_throughput(values: list[float]) -> float:
-    """Return robust median throughput across trial values."""
+    """Compatibility wrapper for suite-level monkeypatching in tests."""
 
-    finite = [float(value) for value in values if math.isfinite(float(value))]
-    if not finite:
-        return 0.0
-    return float(statistics.median(finite))
-
-
-def _build_guardrail_issue(
-    *,
-    metric: str,
-    severity: str,
-    current: float | None,
-    baseline: float | None,
-    degradation_pct: float | None,
-    detail: str,
-) -> dict[str, Any]:
-    """Create a normalized guardrail issue payload."""
-
-    return {
-        "metric": metric,
-        "severity": severity,
-        "current": current,
-        "baseline": baseline,
-        "degradation_pct": degradation_pct,
-        "detail": detail,
-    }
+    return _guardrails._median_throughput(values)
 
 
 def _collect_lineage_guardrails(
@@ -487,380 +374,20 @@ def _collect_lineage_guardrails(
     warn_threshold_pct: float,
     fail_threshold_pct: float,
 ) -> dict[str, Any]:
-    """Measure lineage export overhead and convert it into guardrail metrics."""
+    """Compatibility wrapper preserving suite-level lineage patch points."""
 
-    sample_n = _lineage_guardrail_sample_count(config, suite=suite, num_datasets=num_datasets)
-    if sample_n <= 0:
-        return {"enabled": False}
-
-    sample_seed = int(config.seed + LINEAGE_GUARDRAIL_SEED_OFFSET)
-    with tempfile.TemporaryDirectory(prefix="cauchy_lineage_guardrail_stage_") as tmp_dir:
-        stage_root = Path(tmp_dir)
-        baseline_stage_dir = stage_root / "baseline"
-        current_stage_dir = stage_root / "current"
-        try:
-            bundles_with_lineage, bundles_seen = _stage_lineage_trial_bundles(
-                config,
-                sample_n=sample_n,
-                seed=sample_seed,
-                device=device,
-                baseline_stage_dir=baseline_stage_dir,
-                current_stage_dir=current_stage_dir,
-            )
-        except Exception as exc:
-            return {
-                "enabled": False,
-                "reason": "sampling_failed",
-                "detail": str(exc),
-            }
-        if bundles_seen <= 0:
-            return {"enabled": False}
-
-        lineage_coverage_rate = float(bundles_with_lineage) / float(bundles_seen)
-
-        try:
-            baseline_trials, current_trials = _measure_lineage_persistence_trials(
-                baseline_stage_dir=baseline_stage_dir,
-                current_stage_dir=current_stage_dir,
-                num_bundles=bundles_seen,
-                config=config,
-                trials=LINEAGE_GUARDRAIL_RUNTIME_TRIALS,
-            )
-        except Exception as exc:
-            return {
-                "enabled": False,
-                "reason": "unavailable",
-                "detail": str(exc),
-            }
-    baseline_dpm = _median_throughput(baseline_trials)
-    current_dpm = _median_throughput(current_trials)
-
-    runtime_degradation = degradation_percent("datasets_per_minute", current_dpm, baseline_dpm)
-    runtime_degradation_value = (
-        float(runtime_degradation) if runtime_degradation is not None else 0.0
+    return _guardrails._collect_lineage_guardrails_impl(
+        config,
+        suite=suite,
+        num_datasets=num_datasets,
+        device=device,
+        warn_threshold_pct=warn_threshold_pct,
+        fail_threshold_pct=fail_threshold_pct,
+        sample_count_fn=_lineage_guardrail_sample_count,
+        stage_lineage_trial_bundles_fn=_stage_lineage_trial_bundles,
+        measure_lineage_persistence_trials_fn=_measure_lineage_persistence_trials,
+        median_throughput_fn=_median_throughput,
     )
-    runtime_severity = _severity_from_thresholds(
-        runtime_degradation_value,
-        warn=float(warn_threshold_pct),
-        fail=float(fail_threshold_pct),
-    )
-    runtime_gating_enabled = sample_n >= LINEAGE_GUARDRAIL_RUNTIME_GATING_MIN_SAMPLE
-
-    issues: list[dict[str, Any]] = []
-    if bundles_with_lineage != bundles_seen:
-        issues.append(
-            _build_guardrail_issue(
-                metric="lineage_metadata_coverage",
-                severity="fail",
-                current=float(lineage_coverage_rate),
-                baseline=1.0,
-                degradation_pct=float(max(0.0, (1.0 - lineage_coverage_rate) * 100.0)),
-                detail="Lineage metadata must be present for all generated bundles.",
-            )
-        )
-    if runtime_gating_enabled and runtime_severity != "pass":
-        issues.append(
-            _build_guardrail_issue(
-                metric="lineage_export_runtime_degradation_pct",
-                severity=runtime_severity,
-                current=float(current_dpm),
-                baseline=float(baseline_dpm),
-                degradation_pct=float(runtime_degradation_value),
-                detail=(
-                    "Lineage-enabled shard persistence throughput regressed versus a "
-                    "lineage-stripped control persistence run."
-                ),
-            )
-        )
-
-    return {
-        "enabled": True,
-        "sample_datasets": int(sample_n),
-        "runtime_gating_enabled": bool(runtime_gating_enabled),
-        "runtime_gating_min_sample_datasets": int(LINEAGE_GUARDRAIL_RUNTIME_GATING_MIN_SAMPLE),
-        "runtime_gating_suppressed_reason": (
-            None if runtime_gating_enabled else "insufficient_sample_size"
-        ),
-        "runtime_trials": int(max(1, LINEAGE_GUARDRAIL_RUNTIME_TRIALS)),
-        "runtime_baseline_trials_dpm": baseline_trials,
-        "runtime_with_lineage_trials_dpm": current_trials,
-        "lineage_metadata_coverage_rate": float(lineage_coverage_rate),
-        "runtime_baseline_datasets_per_minute": float(baseline_dpm),
-        "runtime_with_lineage_datasets_per_minute": float(current_dpm),
-        "runtime_degradation_pct": (
-            float(runtime_degradation) if runtime_degradation is not None else None
-        ),
-        "runtime_warn_threshold_pct": float(warn_threshold_pct),
-        "runtime_fail_threshold_pct": float(fail_threshold_pct),
-        "issues": issues,
-        "status": _status_from_issues(issues),
-    }
-
-
-def _matrix_cell_count(matrix: Any) -> int:
-    """Return cell count for a rank-2 matrix-like payload."""
-
-    shape = getattr(matrix, "shape", None)
-    if shape is None or len(shape) < 2:
-        return 0
-    try:
-        n_rows = max(0, int(shape[0]))
-        n_cols = max(0, int(shape[1]))
-    except (TypeError, ValueError):
-        return 0
-    return n_rows * n_cols
-
-
-def _severity_from_thresholds(value: float, *, warn: float, fail: float) -> str:
-    """Map a numeric value to pass/warn/fail severity."""
-
-    if value >= fail:
-        return "fail"
-    if value >= warn:
-        return "warn"
-    return "pass"
-
-
-def _status_from_issues(issues: list[dict[str, Any]]) -> str:
-    """Collapse per-metric issues into an overall status."""
-
-    severities = {str(issue.get("severity", "pass")) for issue in issues}
-    if "fail" in severities:
-        return "fail"
-    if "warn" in severities:
-        return "warn"
-    return "pass"
-
-
-def _resolve_curriculum_guardrail_context(config: GeneratorConfig) -> tuple[bool, str, int | None]:
-    """Resolve curriculum guardrail mode from config."""
-
-    normalized_mode = normalize_curriculum_stage(config.curriculum_stage)
-    if normalized_mode == CURRICULUM_STAGE_OFF:
-        return (False, "off", None)
-    if normalized_mode == CURRICULUM_STAGE_AUTO:
-        return (True, CURRICULUM_STAGE_AUTO, None)
-    return (True, "fixed", int(normalized_mode))
-
-
-@dataclass(slots=True)
-class _CurriculumMetadataCollector:
-    """Collect staged curriculum metadata coverage during benchmark throughput runs."""
-
-    expected_mode: str
-    bundles_seen: int = 0
-    bundles_with_stage_metadata: int = 0
-    mode_mismatch_bundles: int = 0
-
-    def update(self, bundle: DatasetBundle) -> None:
-        """Collect curriculum metadata counters for one generated bundle."""
-
-        self.bundles_seen += 1
-        payload = bundle.metadata.get("curriculum")
-        if not isinstance(payload, dict):
-            return
-        mode_value = payload.get("mode")
-        if mode_value != self.expected_mode:
-            self.mode_mismatch_bundles += 1
-        stage_value = payload.get("stage")
-        if isinstance(stage_value, bool):
-            return
-        if isinstance(stage_value, int) and stage_value in (1, 2, 3):
-            self.bundles_with_stage_metadata += 1
-
-    def build_summary(self) -> dict[str, Any]:
-        """Build curriculum metadata guardrail metrics and issues."""
-
-        coverage_rate = (
-            float(self.bundles_with_stage_metadata) / float(self.bundles_seen)
-            if self.bundles_seen > 0
-            else 0.0
-        )
-
-        issues: list[dict[str, Any]] = []
-        if self.bundles_with_stage_metadata != self.bundles_seen:
-            issues.append(
-                {
-                    "metric": "curriculum_stage_metadata_coverage",
-                    "severity": "fail",
-                    "current": float(coverage_rate),
-                    "baseline": 1.0,
-                    "degradation_pct": float(max(0.0, (1.0 - coverage_rate) * 100.0)),
-                    "detail": "Curriculum stage metadata must be present for all staged bundles.",
-                }
-            )
-        if self.mode_mismatch_bundles > 0:
-            mismatch_rate = float(self.mode_mismatch_bundles) / float(self.bundles_seen)
-            issues.append(
-                {
-                    "metric": "curriculum_mode_metadata_mismatch",
-                    "severity": "fail",
-                    "current": float(mismatch_rate),
-                    "baseline": 0.0,
-                    "degradation_pct": float(mismatch_rate * 100.0),
-                    "detail": "Curriculum metadata mode must match configured staged mode.",
-                }
-            )
-
-        return {
-            "stage_metadata_coverage_rate": float(coverage_rate),
-            "mode_mismatch_bundles": int(self.mode_mismatch_bundles),
-            "issues": issues,
-            "status": _status_from_issues(issues),
-        }
-
-
-@dataclass(slots=True)
-class _MissingnessAcceptanceCollector:
-    """Collect per-bundle missingness metadata for acceptance guardrails."""
-
-    target_rate: float
-    bundles_seen: int = 0
-    bundles_with_metadata: int = 0
-    missing_cells: int = 0
-    total_cells: int = 0
-
-    def update(self, bundle: DatasetBundle) -> None:
-        """Collect missingness counters for one generated bundle."""
-
-        self.bundles_seen += 1
-        payload = bundle.metadata.get("missingness")
-        if not isinstance(payload, dict):
-            return
-
-        total_cells = _matrix_cell_count(bundle.X_train) + _matrix_cell_count(bundle.X_test)
-        if total_cells <= 0:
-            return
-
-        missing_count_raw = payload.get("missing_count_overall")
-        if isinstance(missing_count_raw, bool) or not isinstance(missing_count_raw, (int, float)):
-            return
-        missing_count = int(max(0, min(total_cells, int(missing_count_raw))))
-
-        self.bundles_with_metadata += 1
-        self.total_cells += total_cells
-        self.missing_cells += missing_count
-
-    def build_summary(self) -> dict[str, Any]:
-        """Build acceptance guardrail metrics and issues."""
-
-        coverage_rate = (
-            float(self.bundles_with_metadata) / float(self.bundles_seen)
-            if self.bundles_seen > 0
-            else 0.0
-        )
-        realized_rate = (
-            float(self.missing_cells) / float(self.total_cells) if self.total_cells > 0 else 0.0
-        )
-        rate_abs_error = abs(realized_rate - float(self.target_rate))
-
-        issues: list[dict[str, Any]] = []
-        if self.bundles_with_metadata != self.bundles_seen:
-            issues.append(
-                {
-                    "metric": "missingness_metadata_coverage",
-                    "severity": "fail",
-                    "current": float(coverage_rate),
-                    "baseline": 1.0,
-                    "degradation_pct": float(max(0.0, (1.0 - coverage_rate) * 100.0)),
-                    "detail": "Missingness metadata must be present for all generated bundles.",
-                }
-            )
-
-        rate_error_pp = rate_abs_error * 100.0
-        rate_severity = _severity_from_thresholds(
-            rate_abs_error,
-            warn=MISSINGNESS_RATE_WARN_ABS_ERROR,
-            fail=MISSINGNESS_RATE_FAIL_ABS_ERROR,
-        )
-        if rate_severity != "pass":
-            threshold_pp = (
-                MISSINGNESS_RATE_FAIL_ABS_ERROR * 100.0
-                if rate_severity == "fail"
-                else MISSINGNESS_RATE_WARN_ABS_ERROR * 100.0
-            )
-            issues.append(
-                {
-                    "metric": "missingness_realized_rate_error_pp",
-                    "severity": rate_severity,
-                    "current": float(rate_error_pp),
-                    "baseline": float(threshold_pp),
-                    "degradation_pct": float(rate_error_pp),
-                    "detail": "Realized missing rate drifted from configured target.",
-                }
-            )
-
-        return {
-            "metadata_coverage_rate": float(coverage_rate),
-            "realized_rate_overall": float(realized_rate),
-            "rate_abs_error": float(rate_abs_error),
-            "issues": issues,
-            "status": _status_from_issues(issues),
-        }
-
-
-def _compose_bundle_callback(
-    *,
-    diagnostics_aggregator: CoverageAggregator | None,
-    missingness_acceptance: _MissingnessAcceptanceCollector | None,
-    curriculum_metadata: _CurriculumMetadataCollector | None,
-) -> Callable[[DatasetBundle], None] | None:
-    """Compose optional per-bundle collectors into one callback."""
-
-    if (
-        diagnostics_aggregator is None
-        and missingness_acceptance is None
-        and curriculum_metadata is None
-    ):
-        return None
-
-    def _on_bundle(bundle: DatasetBundle) -> None:
-        if diagnostics_aggregator is not None:
-            diagnostics_aggregator.update_bundle(bundle)
-        if missingness_acceptance is not None:
-            missingness_acceptance.update(bundle)
-        if curriculum_metadata is not None:
-            curriculum_metadata.update(bundle)
-
-    return _on_bundle
-
-
-def _issue_sort_key(issue: dict[str, Any]) -> tuple[int, float]:
-    """Sort regression issues by severity then descending degradation percentage."""
-
-    severity = str(issue.get("severity", "warn"))
-    rank = 0 if severity == "fail" else 1 if severity == "warn" else 2
-    raw_degradation = issue.get("degradation_pct")
-    if isinstance(raw_degradation, bool) or not isinstance(raw_degradation, (int, float)):
-        return (rank, 0.0)
-    degradation = float(raw_degradation)
-    if not math.isfinite(degradation):
-        return (rank, 0.0)
-    return (rank, -degradation)
-
-
-def _collect_guardrail_regression_issues(
-    profile_results: list[dict[str, Any]],
-    *,
-    guardrail_key: str,
-) -> list[dict[str, Any]]:
-    """Flatten one guardrail type into regression issue payloads."""
-
-    issues: list[dict[str, Any]] = []
-    for result in profile_results:
-        guardrails = result.get(guardrail_key)
-        if not isinstance(guardrails, dict) or not bool(guardrails.get("enabled")):
-            continue
-        raw_issues = guardrails.get("issues")
-        if not isinstance(raw_issues, list):
-            continue
-        for issue in raw_issues:
-            if not isinstance(issue, dict):
-                continue
-            merged = dict(issue)
-            merged["profile"] = str(result.get("profile_key"))
-            issues.append(merged)
-    return issues
 
 
 def run_profile_benchmark(
@@ -959,8 +486,8 @@ def run_profile_benchmark(
         result["peak_rss_mb"] = max(0.0, _peak_rss_mb() - rss_before)
         if hw.backend == "cuda" and torch.cuda.is_available():
             try:
-                result["peak_cuda_allocated_mb"] = torch.cuda.max_memory_allocated() / (1024.0**2)
-                result["peak_cuda_reserved_mb"] = torch.cuda.max_memory_reserved() / (1024.0**2)
+                result["peak_cuda_allocated_mb"] = torch.cuda.max_memory_allocated() / MIB
+                result["peak_cuda_reserved_mb"] = torch.cuda.max_memory_reserved() / MIB
             except Exception:
                 result["peak_cuda_allocated_mb"] = None
                 result["peak_cuda_reserved_mb"] = None
@@ -976,7 +503,9 @@ def run_profile_benchmark(
         )
 
     if include_micro:
-        result.update(run_microbenchmarks(config, device=requested_device, repeats=3))
+        result.update(
+            run_microbenchmarks(config, device=requested_device, repeats=MICROBENCH_REPEATS)
+        )
 
     if curriculum_enabled and curriculum_metadata is not None:
         # Control-run configs are derived from an already tuned runtime config;
@@ -1156,7 +685,8 @@ def run_profile_benchmark(
     ):
         profile_segment = _sanitize_profile_key(spec.key)
         if diagnostics_occurrence_total > 1:
-            profile_segment = f"{profile_segment}_run{diagnostics_occurrence_index + 1}"
+            run_number = diagnostics_occurrence_index + DIAGNOSTICS_DUPLICATE_PROFILE_SUFFIX_BASE
+            profile_segment = f"{profile_segment}_run{run_number}"
         profile_diagnostics_dir = diagnostics_root_dir / "diagnostics" / profile_segment
         summary = diagnostics_aggregator.build_summary()
         json_path = write_coverage_summary_json(
